requirements.txt
pandas
matplotlib
sqlalchemy
requests
pyyaml
pytest